+++ website/src/Components/Admin/ConfigManager.js
@@ -26,7 +26,7 @@ import {
     tesseractThreshList,
 } from "defaultOcrConfigs";
 
-import ReturnButton from 'Components/FileSystem/Geral/ReturnButton';
+import ReturnButton from 'Components/FileSystem/ReturnButton';
 import ConfirmLeave from 'Components/Notifications/Geral/ConfirmLeave';
 import Notification from 'Components/Notifications/Geral/Notification';
 import ConfirmActionPopup from 'Components/Form/Geral/ConfirmActionPopup';
