+++ website/src/Components/Admin/StorageManager.js
@@ -15,7 +15,7 @@ import CheckRoundedIcon from "@mui/icons-material/CheckRounded";
 
 import { TimePicker } from "@mui/x-date-pickers/TimePicker";
 
-import ReturnButton from 'Components/FileSystem/Geral/ReturnButton';
+import ReturnButton from 'Components/FileSystem/ReturnButton';
 import Notification from 'Components/Notifications/Geral/Notification';
 import ConfirmActionPopup from 'Components/Form/Geral/ConfirmActionPopup';
 import CheckboxList from 'Components/Form/Geral/CheckboxList';
